backend/requirements.txt
fastapi[all]==0.111.0
uvicorn[standard]==0.29.0
pipecat-ai==0.0.75
google-generativeai==0.5.3
websockets==12.0
prometheus-fastapi-instrumentator==7.1.0
prometheus-client==0.20.0
python-dotenv==1.0.1